+++ emcc.py
@@ -29,6 +29,11 @@ import tarfile
 from dataclasses import dataclass
 from enum import Enum, auto, unique
 
+# This assert needs to happen early, before any too-recent python syntax is used.
+# In particular it needs to happen before we import any python file that uses the
+# `match` keyword.
+assert sys.version_info >= (3, 10), f'emscripten requires python 3.10 or above ({sys.executable} {sys.version})'
+
 from tools import (
   building,
   cache,