+++ src/settings.js
@@ -622,15 +622,10 @@ var GL_ENABLE_GET_PROC_ADDRESS = true;
 // [link]
 var JS_MATH = false;
 
-// If set, enables polyfilling for Math.clz32, Math.trunc, Math.imul, Math.fround.
-// [link]
-var POLYFILL_OLD_MATH_FUNCTIONS = false;
-
 // Set this to enable compatibility emulations for old JavaScript engines. This gives you
 // the highest possible probability of the code working everywhere, even in rare old
 // browsers and shell environments. Specifically:
 //
-// - Add polyfilling for Math.clz32, Math.trunc, Math.imul, Math.fround. (-sPOLYFILL_OLD_MATH_FUNCTIONS)
 // - Disable WebAssembly. (Must be paired with -sWASM=0)
 // - Adjusts MIN_X_VERSION settings to 0 to include support for all browser versions.
 // - Avoid TypedArray.fill, if necessary, in zeroMemory utility function.
@@ -788,28 +783,6 @@ var EXCEPTION_STACK_TRACES = false;
 // https://github.com/WebAssembly/exception-handling/blob/main/proposals/exception-handling/Exceptions.md
 // [compile+link]
 var WASM_LEGACY_EXCEPTIONS = true;
-
-// Emscripten throws an ExitStatus exception to unwind when exit() is called.
-// Without this setting enabled this can show up as a top level unhandled
-// exception.
-//
-// With this setting enabled a global uncaughtException handler is used to
-// catch and handle ExitStatus exceptions.  However, this means all other
-// uncaught exceptions are also caught and re-thrown, which is not always
-// desirable.
-//
-// [link]
-var NODEJS_CATCH_EXIT = false;
-
-// Catch unhandled rejections in node. This only affects versions of node older
-// than 15.  Without this, old version node will print a warning, but exit
-// with a zero return code.  With this setting enabled, we handle any unhandled
-// rejection and throw an exception, which will cause the process to exit
-// immediately with a non-0 return code.
-// This is not needed in Node 15+ so this setting will default to false if
-// MIN_NODE_VERSION is 150000 or above.
-// [link]
-var NODEJS_CATCH_REJECTION = true;
 
 // Whether to support async operations in the compiled code. This makes it
 // possible to call JS functions from synchronous-looking code in C/C++.